Dodan pocetak Ant klase + manji updatei
Dodan pocetak Ant klase s mogucnosti vjerojatnosnoe orijentacijoe po grafu (bazirane na količini feromona). Također dodana mogučnost ispisa odabranog puta.

Svi float-ovi prebačeni u double-ove.

Dodan inline ispis za truth table pri unosu formule.

+++ headers/Graph.h
@@ -11,14 +11,14 @@ public:
     Node* nextZero;
     Node* nextOne;
     Node* sibling;
-    float nextZeroPheromone;
-    float nextOnePheromone;
+    double nextZeroPheromone;
+    double nextOnePheromone;
     bool end;
     bool type;
     int id;
     explicit Node(bool t = NULL, bool e = false);
     void setNodes(Node* nZ, Node* nO);
-    void setPheromones(float nZP, float nOP);
+    void setPheromones(double nZP, double nOP);
 
     ~Node();
 };
@@ -29,7 +29,7 @@ public:
     Node* startNode;
 
     int numberOfLayers;
-    void setPheromone(int fromLayer, bool fromType, bool toType, float value);
+    void setPheromone(int fromLayer, bool fromType, bool toType, double value);
     void printGraph();
     void drawGraph();
     Graph();
